+++ app/public/js/analytics.js
@@ -1,14 +1,13 @@
-(function(global) {
-  'use strict'
-  var $ = global.jQuery;
+((global) => {
+  const $ = global.jQuery;
 
   function trackClick(element, arg1, arg2) {
-    Webtrends.multiTrack({ element: element, argsa: [arg1, arg2, 'WT.dl', '121'] });
+    Webtrends.multiTrack({ argsa: [arg1, arg2, 'WT.dl', '121'], element });
   }
 
-  var trackingName = 'DCSext.IAPT';
+  const trackingName = 'DCSext.IAPT';
 
-  $('.start_button_tracking').on('click', function () {
+  $('.start_button_tracking').on('click', () => {
     trackClick($(this), trackingName, 'StartButton');
   });
 })(window);