+++ app/public/js/init.js
@@ -4,4 +4,4 @@ require('./cookieMessage');
 require('./headerMobileMenuToggle');
 require('./headerMobileSearchToggle');
 require('./headerSearchTypeahead');
-require('./skiplinkFocus');
+require('./querySearchTypeahead');