+++ config/routes.js
@@ -1,6 +1,6 @@
 module.exports = {
   check: { path: '/check', title: 'What happens after you refer yourself' },
   results: { path: '/results' },
-  search: { path: '/search', title: 'We need to know your GP to get you to the right service' },
+  search: { path: '/search', title: 'We need to know your GP surgery to get you to the right service' },
   start: { path: '/', title: 'Find help with stress, anxiety or depression' },
 };