+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+0.2.0 / TBD
+==================
+- Add typeahead for GP search
+- Encode queries for back links
+- Improve accessibility
+- Content and design tweaks
+
 0.1.1 / 2018-07-10
 ==================
 - Rename frontends for consistency